feat(baseline): حفظ الحالة الأولية للمشروع
Refs: production-readiness-report
