+++ package.json
@@ -33,6 +33,7 @@
     "eslint": "^9.0.0",
     "eslint-plugin-react-hooks": "^5.0.0",
     "eslint-plugin-react-refresh": "^0.4.0",
+    "globals": "^16.4.0",
     "jsdom": "^25.0.0",
     "postcss": "^8.5.6",
     "rollup-plugin-visualizer": "^5.12.0",
