+++ src/config/environment.ts
@@ -16,7 +16,7 @@ export interface EnvironmentConfig {
 const getGeminiApiKey = (): string => {
   const apiKey = import.meta.env.VITE_GEMINI_API_KEY;
   if (!apiKey) {
-    throw new Error('GEMINI_API_KEY is not set in the environment variables.');
+    return '';
   }
   return apiKey;
 };
