+++ src/services/test-instructions-system.ts
@@ -40,7 +40,6 @@ export class InstructionsSystemTester {
    * اختبار شامل لنظام التعليمات
    */
   async runFullTest(): Promise<TestResults> {
-    console.log('🧪 بدء اختبار نظام تعليمات الوكلاء...');
     
     const results: TestResults = {
       totalAgents: EXPECTED_AGENTS.length,
@@ -54,29 +53,24 @@ export class InstructionsSystemTester {
     // اختبار تحميل كل وكيل
     for (const agentId of EXPECTED_AGENTS) {
       try {
-        console.log(`📋 اختبار تحميل وكيل: ${agentId}`);
         const instructions = await agentInstructions.getInstructions(agentId);
         
         if (this.validateInstructions(instructions, agentId)) {
           results.successfulLoads++;
           results.loadedAgents.push(agentId);
-          console.log(`✅ نجح تحميل ${agentId}`);
         } else {
           results.failedLoads++;
           results.errors.push(`تعليمات غير صالحة للوكيل: ${agentId}`);
-          console.log(`❌ تعليمات غير صالحة: ${agentId}`);
         }
       } catch (error) {
         results.failedLoads++;
         results.missingAgents.push(agentId);
         results.errors.push(`فشل تحميل ${agentId}: ${error}`);
-        console.log(`❌ فشل تحميل ${agentId}:`, error);
       }
     }
 
     // اختبار حالة التخزين المؤقت
     const cacheStatus = agentInstructions.getCacheStatus();
-    console.log(`💾 حالة التخزين المؤقت: ${cacheStatus.cached.length} وكيل محمل`);
 
     return results;
   }
@@ -89,12 +83,10 @@ export class InstructionsSystemTester {
     
     // التحقق من الحقول الأساسية
     if (!instructions.systemPrompt || typeof instructions.systemPrompt !== 'string') {
-      console.warn(`⚠️ systemPrompt مفقود أو غير صالح للوكيل ${agentId}`);
       return false;
     }
 
     if (!instructions.instructions || !Array.isArray(instructions.instructions)) {
-      console.warn(`⚠️ instructions مفقود أو غير صالح للوكيل ${agentId}`);
       return false;
     }
 
@@ -117,25 +109,13 @@ export class InstructionsSystemTester {
    * طباعة تقرير مفصل
    */
   printReport(results: TestResults): void {
-    console.log('\n📊 تقرير اختبار نظام التعليمات:');
-    console.log('================================');
-    console.log(`📈 إجمالي الوكلاء: ${results.totalAgents}`);
-    console.log(`✅ تم تحميلها بنجاح: ${results.successfulLoads}`);
-    console.log(`❌ فشل التحميل: ${results.failedLoads}`);
-    console.log(`📊 معدل النجاح: ${((results.successfulLoads / results.totalAgents) * 100).toFixed(1)}%`);
     
     if (results.missingAgents.length > 0) {
-      console.log('\n🚫 الوكلاء المفقودون:');
-      results.missingAgents.forEach(agent => console.log(`  - ${agent}`));
     }
     
     if (results.errors.length > 0) {
-      console.log('\n⚠️ الأخطاء:');
-      results.errors.forEach(error => console.log(`  - ${error}`));
     }
     
-    console.log('\n✅ الوكلاء المحملون بنجاح:');
-    results.loadedAgents.forEach(agent => console.log(`  - ${agent}`));
   }
 }
