+++ src/tests/comprehensive-basmala-test.cjs
@@ -25,7 +25,7 @@ class ScreenplayClassifier {
   }
 
   static hasSentencePunctuation(s) {
-    return /[\.!\؟\?]/.test(s);
+    return /[.!؟?]/.test(s);
   }
 
   static wordCount(s) {
