+++ verification/p0-status.txt
@@ -1,0 +1,1 @@
+✅ P0 fixes verified successfully
